Modified according to PR comments

+++ bpf/types.h
@@ -161,7 +161,7 @@ typedef struct pkt_info_t {
 // Structure for payload metadata
 typedef struct payload_meta_t {
     u32 if_index;
-    u16 pkt_len;
+    u32 pkt_len;
     u64 timestamp;  // timestamp when packet received by ebpf
 } __attribute__((packed)) payload_meta;
